Redefine string_t. Fix wrong variables type.

+++ dlls/util.h
@@ -37,13 +37,13 @@ extern globalvars_t				*gpGlobals;
 #define STRING(offset)		(const char *)(gpGlobals->pStringBase + (int)offset)
 
 #if !defined XASH_64BIT || defined(CLIENT_DLL)
-#define MAKE_STRING(str)	((size_t)str - (size_t)STRING(0))
+#define MAKE_STRING(str)	((int)str - (int)STRING(0))
 #else
 static inline int MAKE_STRING(const char *szValue)
 {
 	long long ptrdiff = szValue - STRING(0);
 	if( ptrdiff > INT_MAX || ptrdiff < INT_MIN )
-		return ALLOC_STRING(szValue);
+		return ALLOC_STRING( szValue );
 	else
 		return (int)ptrdiff;
 }
